geolocationService/requirements.txt
fastapi==0.115.3
uvicorn[standart]==0.32.0
sqlalchemy[asyncio]==2.0.36
pydantic-settings==2.6.0
asyncpg==0.30.0
psycopg2==2.9.10
alembic==1.13.3
websockets==13.1
httpx==0.27.2
pyjwt[crypto]==2.9.0
